requirements.txt
streamlit
torch
torchvision
faiss-cpu
numpy
pillow
joblib
requests